fix: clear the last clang-tidy findings from tests, tools and examples
Byte caps written `1024 * 1024` were computed in int and widened to std::size_t
only afterwards; they are widened first now. Powers of two destined for a signed
result are built in uint64 and converted. Loop counters used as bit patterns or
shift counts are declared unsigned, which is what they were. A power-of-two test
(`u & (u - 1)`) moves to the unsigned domain the trick is stated for, and char
is cast through unsigned char before being xored, because plain char is signed
here.

Two narrowing conversions are made explicit: an iterator distance is a signed
difference_type, and a uint64 round count reaching a double multiply says so.
One nested if is merged by hand.

The suite's single getenv keeps a NOLINT rather than a rewrite: it reads the
Postgres DSN once, on the main thread, before the file starts any thread of its
own, and nothing in the suite calls setenv.

The repository now reports zero clang-tidy findings under its own configuration,
against 2,645 when the first whole-repository sweep became possible. The 442 that
remain are in the vendored fastestjsoninthewest submodule, which carries its own
.clang-tidy and is not ours to edit.

+++ tests/feature_polycalc_tests.cpp
@@ -411,7 +411,8 @@ auto main() -> int {
                       for (std::int64_t k = 0; k <= n; ++k) {
                           sum += binomial(n, k).value();
                       }
-                      t.expect(sum == (std::int64_t{1} << n), "sum_k C(n,k) == 2^n");
+                      t.expect(sum == static_cast<std::int64_t>(std::uint64_t{1} << static_cast<unsigned>(n)),
+                           "sum_k C(n,k) == 2^n");
                   }
               })
         .test("combinatorics_factorial_sequences",
